disable yuv recon
Emergency measure, because the MKV output patch adds bits which
cause compilation failures in YUVOutput starting with the
changes made in ded33e446.  The better solution would be fixing
the real source of the error, but for the moment just make sure
the source builds with the MKV patch included.